Modularize Control Components
1. Tabs => Tab
2. Filter => Separate Filter.js file
3. Fixed eslint warnings
